requirements.txt

python-telegram-bot==20.3
gspread
oauth2client
openai
rapidfuzz
flask
nest_asyncio
